Creating new files for the shared networking code
