+++ Tests/ItemProviderTests.swift
@@ -652,41 +652,3 @@ class ItemProviderTests: XCTestCase {
         wait(for: [expectation], timeout: 2)
     }
 }
-
-struct TestProviderRequest: ProviderRequest {
-    
-    let persistenceKey: Key?
-    var baseURL: URL { URL(string: "https://www.google.com")! }
-    var path: String { "" }
-
-    init(key: Key = "TestExample") {
-        self.persistenceKey = key
-    }
-}
-
-struct TestPostProviderRequest: ProviderRequest {
-    let persistenceKey: Key?
-    var baseURL: URL { URL(string: "https://www.google.com")! }
-    var path: String { "" }
-
-    init(key: Key = "TestPostExample") {
-        self.persistenceKey = key
-    }
-    
-    var httpMethod: HTTPMethod {
-        return .post
-    }
-}
-
-struct TestItem: Providable {
-    var identifier: Key { return title }
-    
-    let title: String
-}
-
-extension FileManager {
-    
-    var cachesDirectoryURL: URL! { //swiftlint:disable:this implicitly_unwrapped_optional
-        return urls(for: .cachesDirectory, in: .userDomainMask).first
-    }
-}
