+++ Tests/TestPostProviderRequest.swift
@@ -1,0 +1,25 @@
+//
+//  TestPostProviderRequest.swift
+//  ProviderTests
+//
+//  Created by Ashli Rankin on 1/16/24.
+//  Copyright © 2024 Lickability. All rights reserved.
+//
+
+import Foundation
+import Provider
+import Networking
+
+struct TestPostProviderRequest: ProviderRequest {
+    let persistenceKey: Key?
+    var baseURL: URL { URL(string: "https://www.google.com")! }
+    var path: String { "" }
+
+    init(key: Key = "TestPostExample") {
+        self.persistenceKey = key
+    }
+    
+    var httpMethod: HTTPMethod {
+        return .post
+    }
+}
